Small leak in globus_get_cert_and_chain()
Valgrind shows a leak of 24 bytes per GSI authentication in globus_get_cert_and_chain(). It appears to be caused by BIO_NOCLOSE. The BIO is allocated with BIO_new_mem_buf(), so the data is treated as an unmanaged, read-only buffer, and there's no need to set the BIO_NOCLOSE flag.