Add no-authz mode which skips LCMAPS.
If --no-authz is passed to the Xrootd-LCMAPS module, then it will
skip the invocation of LCMAPS, meaning the username will not be filled
out.  Instead, the name will be inherited from Xrootd - either as the
DN or the OpenSSL subject hash.

With this mode, LCMAPS does not need to be setup on the host, nor does
a Unix user have to exist for each Xrootd mapping.