Copy the authz key into the monitoring info.
Some GSI modes map the DN hash to the `XrdSecEntity.name` structure,
meaning that our monitoring no longer receives the actual DN - a
significant decrease in capabilities.

This commit causes the LCMAPS plugin to always copy the authz key
(a concatenation of the DN and VOMs attributes) into the moninfo;
the monitoring system can fallback to utilizing this info when the
DN hash is reported as the name.